Allow solidus_support 0.4
In stores that use e.g. the solidus_paypal_braintree extension which is
in a state right now that makes maintenance hard, we cannot upgrade to
the latest master of this extension because of this gem depending on
`solidus_support` `~> 0.5`.

I tested using 0.4 using the Gemfile, and it does work.